37 Hour
Started work on reimplementing the drawing system, now in its own scene, now with a toggle for grid or formless drawing! Still debating on which i want to do, but odds are i hook it up to a button. Difficulty now shows on the cool car spinny screen i dont have name for