42Hour 42Minute
This commit is for randomness week day 4. I lost track of time a bit and did 2 hours and 10 minutes. Being synced to the hour is a big mess in my project. I decided to give up on switching from 3d scene to 2d scene and just put the result screen there til i figure things out